refactor(components/atom/popover): lint
