+++ components/atom/popover/demo/index.js
@@ -1,7 +1,5 @@
 /* eslint-disable react/prop-types, no-unused-vars, no-console */
 
-import {forwardRef} from 'react'
-
 import {Anchor, H1, Paragraph} from '@s-ui/documentation-library'
 
 import ArticleArrow from './articles/ArticleArrow.js'
@@ -11,22 +9,10 @@ import ArticleDefault from './articles/ArticleDefault.js'
 import ArticlePosition from './articles/ArticlePosition.js'
 import ArticleRef from './articles/ArticleRef.js'
 import ArticleType from './articles/ArticleType.js'
+import Content from './Content.js'
+import {CLASS_SECTION} from './settings.js'
 
 import './index.scss'
-
-const BASE_CLASS_DEMO = `DemoAtomPopover`
-const CLASS_SECTION = `${BASE_CLASS_DEMO}-section`
-
-const Content = forwardRef((props, forwardedRef) => {
-  return (
-    <div ref={forwardedRef} style={{width: '200px', padding: '8px'}}>
-      <span {...props}>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut id mauris
-        ornare, imperdiet nunc a, interdum dolor.
-      </span>
-    </div>
-  )
-})
 
 const Demo = () => {
   return (
