+++ components/atom/popover/demo/settings.js
@@ -1,0 +1,2 @@
+const BASE_CLASS_DEMO = `DemoAtomPopover`
+export const CLASS_SECTION = `${BASE_CLASS_DEMO}-section`
